bifurcate 'network actors' tests.
